+++ src/components/Toggle/Toggle.test.tsx
@@ -3,6 +3,7 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { axe } from "jest-axe";
 import renderer from "react-test-renderer";
+
 import * as generateUUID from "../../helpers/generateUUID";
 import Toggle from "./Toggle";