+++ src/components/Toggle/ToggleTypes.tsx
@@ -1,4 +1,4 @@
 export enum ToggleSizes {
-  Large = "large",
   Small = "small",
+  Large = "large",
 }