+++ src/theme/components/notification.ts
@@ -2,7 +2,7 @@ import { NotificationTypes } from "../../components/Notification/NotificationTyp
 
 const Notification = {
   parts: ["container", "dismissibleButton", "icon"],
-  baseStyle: ({ centered, dismissible, noMargin, notificationType }) => {
+  baseStyle: ({ dismissible, isCentered, noMargin, notificationType }) => {
     let bg = "ui.status.primary";
     if (
       notificationType === NotificationTypes.Announcement ||
@@ -15,7 +15,7 @@ const Notification = {
       display: "flex",
       fontSize: "-1", // slightly smaller than the default size
       position: "relative",
-      textAlign: centered ? "center" : null,
+      textAlign: isCentered ? "center" : null,
       borderRadius: noMargin ? "0" : "4px",
       margin: noMargin ? "0" : "s",
       container: {
@@ -23,7 +23,7 @@ const Notification = {
         maxWidth: "var(--nypl-breakpoint-xl)",
         padding: "s",
         paddingRight: dismissible ? "l" : null,
-        paddingLeft: centered && dismissible ? "l" : null,
+        paddingLeft: isCentered && dismissible ? "l" : null,
         width: "100%",
       },
       dismissibleButton: {
@@ -78,7 +78,7 @@ const NotificationContent = {
 
 const NotificationHeading = {
   parts: ["heading"],
-  baseStyle: ({ centered, icon, notificationType }) => {
+  baseStyle: ({ icon, isCentered, notificationType }) => {
     let color = "ui.black";
     if (notificationType === NotificationTypes.Announcement) {
       color = "section.research.secondary";
@@ -88,7 +88,7 @@ const NotificationHeading = {
     return {
       display: "flex",
       marginBottom: "xxs",
-      justifyContent: centered ? "center" : null,
+      justifyContent: isCentered ? "center" : null,
       heading: {
         marginBottom: "0",
         marginTop: icon ? "xxxs" : "0",